requirements.txt
playwright
gspread
oauth2client
requests
python-dotenv